Fix linter options and linter errors

+++ pkg/products/inventory/inventory_test.go
@@ -44,8 +44,7 @@ func TestInventory_GetProductStock(t *testing.T) {
 	itemStock, err := productInventory.GetProductStock(itemID)
 
 	g.Expect(err).ShouldNot(HaveOccurred())
-	g.Expect(itemStock).To(Equal(uint(stock)))
-
+	g.Expect(itemStock).To(Equal(stock))
 }
 
 func TestInventory_GetProductStock_Error(t *testing.T) {
@@ -85,7 +84,6 @@ func TestInventory_HasInStock(t *testing.T) {
 
 	g.Expect(err).ShouldNot(HaveOccurred())
 	g.Expect(has).To(BeTrue())
-
 }
 
 func TestInventory_HasInStock_False(t *testing.T) {
@@ -145,7 +143,6 @@ func TestInventory_GetPrice(t *testing.T) {
 
 	g.Expect(err).ShouldNot(HaveOccurred())
 	g.Expect(productPrice).To(Equal(price))
-
 }
 
 func TestInventory_GetPrice_Error(t *testing.T) {
